refactor a little and make existing test pass
