+++ package.json
@@ -14,6 +14,7 @@
     "watch-test": "make watch-test"
   },
   "devDependencies": {
+    "admin-on-rest": "^0.8.1",
     "babel-cli": "6.22.2",
     "babel-core": "6.22.1",
     "babel-plugin-istanbul": "4.0.0",
@@ -33,6 +34,7 @@
     "graphql-tag": "1.2.4",
     "isomorphic-fetch": "2.2.1",
     "lodash.merge": "4.6.0",
+    "lodash.omit": "4.5.0",
     "lodash.pickby": "4.6.0",
     "pluralize": "3.1.0",
     "react-router": "3.0.2",
@@ -40,6 +42,6 @@
     "redux-saga": "0.14.3"
   },
   "peerDependencies": {
-    "admin-on-rest": "~0.8.0"
+    "admin-on-rest": "^0.8.1"
   }
 }
