+++ src/realtime/queryObserver.js
@@ -1,0 +1,13 @@
+import { END } from 'redux-saga';
+
+export default emitter => ({
+    complete() {
+        emitter(END);
+    },
+    error() {
+        emitter(END);
+    },
+    next(apolloQueryResult) {
+        emitter(apolloQueryResult);
+    },
+});
